requirements.txt
python-dotenv
tweepy
pymongo
